[refactor] move files around in store module
