+++ nearcore/src/lib.rs
@@ -38,8 +38,8 @@ use near_epoch_manager::shard_tracker::{ShardTracker, TrackedConfig};
 use near_network::PeerManagerActor;
 use near_primitives::genesis::GenesisId;
 use near_primitives::types::EpochId;
+use near_store::db::metadata::DbKind;
 use near_store::genesis::initialize_sharded_genesis_state;
-use near_store::metadata::DbKind;
 use near_store::metrics::spawn_db_metrics_loop;
 use near_store::{NodeStorage, Store, StoreOpenerError};
 use near_telemetry::TelemetryActor;
