+++ tools/protocol-schema-check/src/main.rs
@@ -18,7 +18,6 @@ use near_vm_runner::*;
 
 use near_schema_checker_lib::{FieldName, FieldTypeInfo, ProtocolSchema, ProtocolSchemaInfo};
 use near_stable_hasher::StableHasher;
-use near_store::epoch_info_aggregator::EpochInfoAggregator;
 use std::any::TypeId;
 use std::collections::{BTreeMap, HashSet};
 use std::fs;
