+++ tools/speedy_sync/src/main.rs
@@ -5,6 +5,7 @@ use near_chain::types::{ChainConfig, Tip};
 use near_chain::{Chain, ChainGenesis, DoomslugThresholdMode};
 use near_chain_configs::{GenesisValidationMode, MutableConfigValue, ReshardingConfig};
 use near_epoch_manager::EpochManager;
+use near_epoch_manager::epoch_info_aggregator::EpochInfoAggregator;
 use near_epoch_manager::shard_tracker::{ShardTracker, TrackedConfig};
 use near_primitives::block::Block;
 use near_primitives::block_header::BlockHeader;
@@ -16,7 +17,6 @@ use near_primitives::merkle::PartialMerkleTree;
 use near_primitives::types::EpochId;
 use near_primitives::utils::index_to_bytes;
 use near_store::HEADER_HEAD_KEY;
-use near_store::epoch_info_aggregator::EpochInfoAggregator;
 use near_store::{DBCol, Mode, NodeStorage, Store, StoreUpdate};
 use near_time::Clock;
 use nearcore::{NightshadeRuntime, NightshadeRuntimeExt};
